+++ src/scan.c
@@ -61,7 +61,7 @@ static int get_char(void)
     // FIXME TODO or check for error? feof() vs. ferror()
     fclose(TT.scs->fp);
     TT.scs->fp = 0;
-    TT.scs->p = "  " + 2;
+    TT.scs->p = &("  "[2]);   // Quiet clang warning on "  " + 2;
     if (!TT.scs->prog_args) {
       xfree(TT.scs->line);
       if (lastchar == '\n') return EOF;
@@ -285,7 +285,7 @@ static void ascan_opt_div(int div_op_allowed_here)
       TT.scs->toktype = BUILTIN;
       TT.scs->tok = tkbuiltin;
       TT.scs->tokbuiltin = n;
-    } else if ((TT.scs->ch == '(')) {
+    } else if (TT.scs->ch == '(') {
       TT.scs->toktype = USERFUNC;
       TT.scs->tok = tkfunc;
     } else {